add  delete condition for image, validation and imge files for articles

+++ models/ImageUpload.php
@@ -10,15 +10,36 @@ class ImageUpload extends Model
 {
 	public $image;
 
+	public function rules()
+	{
+		return [
+			[['image'], 'required'],
+			[['image'], 'file', 'extensions' => 'jpg,png']
+		];
+	}
+
+
 	public function uploadFile(UploadedFile $file, $currentImage)
 	{
 		$this->image = $file;
 
-		unlink(Yii::getAlias('@web') . 'uploads/' . $currentImage);
+		if ($this->validate()) {
+			$this->deleteCurrentImage($currentImage);
 
-		$filename = strtolower(md5(uniqid($file->baseName)) . '.' . $file->extension);
-		$file->saveAs(Yii::getAlias('@web') . 'uploads/' . $filename);
+			$filename = strtolower(md5(uniqid($file->baseName)) . '.' . $file->extension);
+			$file->saveAs(Yii::getAlias('@web') . 'uploads/' . $filename);
 
-		return $filename;
+			return $filename;
+		}
+	}
+
+	public function deleteCurrentImage($currentImage)
+	{
+		if (
+			file_exists(Yii::getAlias('@web') . 'uploads/' . $currentImage) &&
+			is_file(Yii::getAlias('@web') . 'uploads/' . $currentImage)
+		) {
+			unlink(Yii::getAlias('@web') . 'uploads/' . $currentImage);
+		}
 	}
 }
